loginphone page refactoring
